changed precision in simplereweighter from float to double, observed problems with large mc weights in mg5_amc@nlo

+++ Analysis/interface/simpleReweighter.h
@@ -20,7 +20,7 @@ public:
 
     void reWeight( float &oldweight);
 
-    float getRenormalization()const;
+    double getRenormalization()const;
 
 
     void switchOff(bool switchoff){switchedoff_=switchoff;}
@@ -29,7 +29,7 @@ public:
 
 private:
     float tmpweight_;
-    float unwcounter_,wcounter_;
+    double unwcounter_,wcounter_;
 protected:
     bool switchedoff_;
 };
