+++ Analysis/src/simpleReweighter.cc
@@ -6,6 +6,7 @@
  */
 
 #include "../interface/simpleReweighter.h"
+#include <iostream>
 
 namespace ztop{
 
@@ -14,15 +15,15 @@ simpleReweighter::simpleReweighter():tmpweight_(1),unwcounter_(0),wcounter_(0),s
 }
 
 void simpleReweighter::reWeight( float &oldweight){
-	if(switchedoff_) return;
-    unwcounter_+=oldweight;
+    if(switchedoff_) return;
+    unwcounter_+= (double) oldweight;
     float newweight=oldweight*tmpweight_;
-    wcounter_+=newweight;
+    wcounter_+=(double) newweight;
     oldweight=newweight;
 }
 
-float simpleReweighter::getRenormalization()const{
-	if(switchedoff_) return 1;
+double simpleReweighter::getRenormalization()const{
+    if(switchedoff_) return 1;
     if(wcounter_!=0)return unwcounter_/wcounter_;
     else return 1;
 }
